+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 # v0.2.1
 
 - Fixed `TxTable.Put()` callable with read only tables
+- Added `Table()` and `MustTable()` to `Tx`
 
 # v0.2.0